requirements.txt
certifi==2018.11.29
chardet==3.0.4
color-transfer==0.1
colormath==3.0.0
decorator==4.3.0
idna==2.8
networkx==2.2
numpy==1.15.4
opencv-python==3.4.4.19
Pillow==5.3.0
requests==2.21.0
urllib3==1.24.1
